Add form id attribute
Add id attribute to <form> inside all single-product add-to-cart templates, for plugins/add-ons which might add <inputs> outside the <form>. Inputs can then use the form="" attribute to be included in the form submission.

ID attribute added to add-to-cart <form> for product pages (simple used as example, variation etc. would also need ID attributes added). This addition could allow for more 3rd party functionality on the product pages, since as it is everything needs to be attached to one of 2-3 hooks inside the <form>.